feat: seguridad, refactor y flujo completo de pedidos
## Cambios generales
- Auditoria de seguridad: todos los endpoints protegidos con validacion
  de token JWT via X-User-Id / X-User-Roles headers
- .gitignore: ignorar .vscode, .venv, target/, *.jar, *.bat, *.sh,
  bases de datos locales, logs, .env y archivos de IDE

## Refactor de controladores
- Eliminado parametro redundante idUsuario en endpoints de carrito,
  pedidos y notificaciones — ahora se lee del header X-User-Id
- CrearPedidoDTO simplificado: solo direccionEntrega + courier

## Correcciones en stock (Fix A, B, C)
- InventarioServiceImpl.buildInventario(): inicializar
  cantidadReservada=0 para evitar DataIntegrityViolationException
- schema.sql (ms-stock): agregada tabla idempotencia_keys que faltaba,
  causaba SQLGrammarException (500) en toda operacion de escritura
- security-service: eliminado seed inseguro de ROLE_ADMIN en schema.sql,
  creado SecurityInitializer con usuario admin configurable

## Sistema de reserva de stock con idempotencia
- Nuevos endpoints: PUT /api/stock/{id}/reservar, confirmar-reserva,
  liberar-reserva con claves deterministicas
- Correccion de doble resta en InventarioServiceImpl.reservarStock()
- FeignClientConfig ya no propaga X-User-Roles (Feign internal bypass)

## Flujo de pedidos unificado (Fix D)
- pagarPedido() ahora ejecuta todo el post-pago: confirmar stock,
  crear envio, vaciar carrito, notificar → EN_PREPARACION directo
- confirmarPedido() mantenido para pedidos legacy en estado PAGADO
- validarTransicion permite CREADO → EN_PREPARACION

## Reintento de pago (Fix E)
- Si ms-pagos rechaza el pago (20% aleatorio simulado), reintenta
  automaticamente con reintentarPago() en lugar de fallar
- Nuevos metodos Feign en PagoClient: consultarPorPedido,
  reintentarPago, consultarPorId con fallbacks
- GUIA-PRUEBAS-E2E.md actualizada con flujo mergeado

+++ ms-catalogo/src/main/java/com/ms_catalogo/dto/VarianteResponseDTO.java
@@ -3,6 +3,7 @@ package com.ms_catalogo.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -12,5 +13,5 @@ public class VarianteResponseDTO {
     private Long idPerfume;
     private String sku;
     private Integer ml;
-    private Double precio;
+    private BigDecimal precio;
 }
