+++ ms-pedidos/src/main/java/com/ms_pedidos/client/dto/VarianteResponseDTO.java
@@ -2,13 +2,14 @@ package com.ms_pedidos.client.dto;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 public class VarianteResponseDTO {
-    private Long    idVariante;
-    private Long    idPerfume;
-    private String  sku;
-    private Integer ml;
-    private Double  precio;
+    private Long      idVariante;
+    private Long      idPerfume;
+    private String    sku;
+    private Integer   ml;
+    private BigDecimal precio;
 }
